moved swagger ui to api endpoints, tidied code.
