+++ orcid-api-common/src/main/resources/org/orcid/api/common/swagger/swagger-ui.ftl
@@ -1,3 +1,21 @@
+<#--
+
+    =============================================================================
+
+    ORCID (R) Open Source
+    http://orcid.org
+
+    Copyright (c) 2012-2014 ORCID, Inc.
+    Licensed under an MIT-Style License (MIT)
+    http://orcid.org/open-source-license
+
+    This copyright and license information (including a link to the full license)
+    shall be included in its entirety in all copies or substantial portion of
+    the software.
+
+    =============================================================================
+
+-->
 <!DOCTYPE html>
 <html>
 <head>
