+++ orcid-persistence/src/main/resources/staging-persistence.properties
@@ -154,9 +154,5 @@ org.orcid.core.token.read_validity_seconds=631138519
 org.orcid.shibboleth.enabled=true
 
 #swagger
-#org.orcid.swagger.memberhost=localhost:8080
-#org.orcid.swagger.memberbasepath=/orcid-api-web
-#org.orcid.swagger.publicbasepath=/orcid-pub-web
-#org.orcid.swagger.publichost=localhost:8080
 org.orcid.swagger.authendpoint=http://localhost:8080/orcid-web/oauth/authorize
 org.orcid.swagger.tokenendpoint=http://localhost:8080/orcid-api-web/oauth/token
