+++ orcid-pub-web/src/main/java/org/orcid/api/t1/server/T1OrcidApiServiceImplV1_2.java
@@ -18,9 +18,6 @@ package org.orcid.api.t1.server;
 
 import javax.ws.rs.Path;
 
-import io.swagger.annotations.Api;
-
-
 /**
  * 
  * @author Will Simpson
@@ -30,4 +27,3 @@ import io.swagger.annotations.Api;
 public class T1OrcidApiServiceImplV1_2 extends T1OrcidApiServiceImplBase {
 
 }
-
